% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{fars_map_state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{is the ID number (integer) for the state for which the user is requesting a map}

\item{year}{is the year (integer) of the data the user wants to see mapped for the requested state}
}
\value{
Function returns a data map of the data for the requested state
}
\description{
This function displays maps of data for the year specified by user for a state specified by the user
}
\note{
This function will fail if an invalid state ID number (state.num), is specified

-- state numbers from 1 t 50 plus, but don't include 2 or 3
}
