% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{fars_read}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{is the name (string) of the file that will be read}
}
\value{
Function returns a table of data from the specified file
}
\description{
Function reads data from an external csv file
}
\note{
This function will fail if a file with the specified name does not exist
}
