% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{fars_read_years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{is the set of years for which the user is requesting data tables or maps}
}
\value{
Function returns a table for each requested year. Each table contains columns for MONTH and year
}
\description{
Function reads the data for the years specified
  importFrom dplyr %>%
}
\note{
This function will fail if an invalid year is specified, if the "years" imply a file that does not exist
}
