% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{is the set of years (integers) for which the user is requesting tables of data}
}
\value{
Function returns a table that contains data for Month and year plus summary lines in the table tabulate the number of entries for each MONTH and year group
}
\description{
This function calls other functions so that it eventually returns a table to the user of data corresponding to the years specified by user
}
