% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{make_filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{is the year of the data to be examined ... year can be integer or string (because input will be converted to integer)}
}
\value{
Function returns a file name that includes the specified year
}
\description{
Function creates the name of the file to be read based on the specified year
}
\examples{
   make_filename(2013)
   make_filename(2015)

}
